// frontend/src/app/core/generated/circabc/model/userActionLog.ts
/**
 * CIRCABC REST API
 *
 * Contact: DIGIT-CIRCABC-SUPPORT@ec.europa.eu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Node } from './node';


export interface UserActionLog { 
    actionDate?: string;
    action?: string;
    node?: Node;
    igNode?: string;
    username?: string;
}